Re-organize a smidge
