+++ tests/tests/full-migration.test.ts
@@ -2,25 +2,26 @@ import { execa } from 'execa';
 import fse from 'fs-extra';
 import { beforeAll, describe, expect, test } from 'vitest';
 
-import { assertEmberTest, extractTests } from './assertions.js';
+import { assertEmberTest, migrate } from '../assertions.js';
 import {
   type Project,
   addonFrom,
   build,
   findFixtures,
+  lintAddon,
   lintTestApp,
-} from './helpers.js';
+} from '../helpers.js';
 
 let fixtures = await findFixtures();
 
-describe('extract-tests: fixtures', () => {
+describe('default command: fixtures', () => {
   for (let fixtureName of fixtures) {
     describe(fixtureName, () => {
       let project: Project;
 
       beforeAll(async () => {
         project = await addonFrom(fixtureName);
-        await extractTests(project, ['--in-place']);
+        await migrate(project);
         await build(project);
       });
 
@@ -32,6 +33,12 @@ describe('extract-tests: fixtures', () => {
         expect(await fse.pathExists(project.testAppPath), 'testAppPath').toBe(
           true
         );
+      });
+
+      test.concurrent('lint addon', async () => {
+        let result = await lintAddon(project);
+
+        expect(result).toMatchObject({ exitCode: 0 });
       });
 
       test.concurrent('lint test-app', async () => {
